Locking in linked list median filter. Refactoring tests.
