+++ debug/timing.js
@@ -116,12 +116,12 @@ singleTests.stdDevSubtract = function() { return stdDevSubtract(medium, medium2)
 */
 
 medium = [3,5,4,2,5,3,6,3,9,5,3,2,4,3,5,3,8,7,7,2,4];
-
+/*
 doubleTests.medianFilter = [
     function() { return medianFilter(im.intensity, quasarVarianceMedian); },
     function() { return medianFilter2(im.intensity, quasarVarianceMedian); }
 ];
-
+*/
